API

Distributions.logpdf — Method

logpdf(m::BetaSplitTree, γ, δ)

Get the log-probability of observing clade γ with split δ under the Beta-splitting model m.

SmoothTree._inner_serial! — Method

_inner_serial!(...)

This is in the ABC simulation algorithm within the EP site update (see Bartelmé & Chopin), implemented in a serial fashion.

SmoothTree._inner_threaded! — Method

_inner_threaded!(...)

See _inner_serial!, but in a multi-threaded implementation using a user-defined batch size for simulation batches executed in parallel.

SmoothTree._momentmatching — Method

_momentmatching(...)

Compute the new global approximation by approximating the tilted distribution within the exponential family using some form of moment matching.

SmoothTree._tuneh! — Method

_tuneh!(...)

Tune the h parameter and re-evaluate accepted simulations.

SmoothTree.ep! — Function

ep!(alg, n; kwargs...)

Do n serial EP passes.

SmoothTree.ep_iteration — Method

ep_iteration(alg, i)

Do a single EP site update.

SmoothTree.ep_serial! — Method

ep_serial!(alg; rnd=true)

Serial EP pass. When rnd=true the pass goes over the data in a random order.

SmoothTree.logpartition — Method

logpartition(model::MSCModel)

Compute the logpartition function of the MSCModel.

SmoothTree.logpartition — Method

logpartition(x::AbstractMBM)

Compute the log-partition function for x.

The log-partition function of a categorical distribution on k categories with moment parameter θ = (θ1, θ2, …, θ{k-1}) is -log(1-∑i θi) = -log θk. The MBM defines a categorical distribution on tree topologies. We have defined an order on trees (in particular we have a well-defined last tree, see reftree). So it appears we can easily compute -log θk.

SmoothTree.matchmoments — Method

updated_model(trees, cavity, α)

Method to update the full approximation by moment matching. This matches the moments of the MBM distribution to the Dirichlet-MBM posterior with prior α for the accepted trees, and updates the Gaussian distributions for the branch parameters.

SmoothTree.pep! — Function

pep!(alg, n)

Do n (embarrassingly) parallel EP passes.

SmoothTree.prune — Method

prune

Prune a sparsely represented MBM object by setting all represented splits with split probabilities indistinguishable from the probability of an unrepresented split to the latter (thereby removing the split from the set of explicitly represented splits).

SmoothTree.prune — Method

prune(x::SparseSplits, atol=1e-9)

Prune out splits which are barely supported (have η ≈ η0, with absolute tolerance atol)

SmoothTree.randsplit — Method

randsplit(m::BetaSplitTree, γ)

Generate a random split of clade γ for the Beta splitting model m

SmoothTree.randsplitofsize — Method

randsplitofsize(γ, k)

Pick a split uniformly from the set of splits of size k in clade γ.

SmoothTree.randtree — Method

randtree(model::MSCSampler)

Simulate a species tree from an MSCModel (in the EP-ABC algorithm this is used to simulate from the cavity)

SmoothTree.traceback — Method

traceback

SmoothTree.BetaSplitTree — Type

BetaSplitTree(β, n)

Stores the relevant pmf's for a Beta-splitting Markov branching model on cladograms.

β ∈ (-2, Inf]: the shape parameter (β → -2.0 gives comb tree with probability one, β=-1.5 gives PDA, β=-1 gives AB, β=0 gives Yule, β=Inf gives random partition tree)

p: Beta-splitting pmfs for clades of size 3 to n

q: Beta-splitting probabilities per split, i.e. probability of a split of size i of a clade of size k is p[k-2][i]/binomial(k,i) when 2i ≠ k.

n: root clade size

References

Aldous, David. "Probability distributions on cladograms." Random discrete structures. Springer, New York, NY, 1996. 1-18.

Jones, Graham R. "Tree models for macroevolution and phylogenetic analysis." Systematic biology 60.6 (2011): 735-746.

SmoothTree.BiMap — Type

BiMap{T,V}

Bijective map for distinct types, so that getindex dispatches on type.

SmoothTree.BranchModel — Type

BranchModel{T,V}

The BranchModel stores for each clade, potentially, a natural parameter vector, and has a fallback η0 representing the natural parameter vector for unrepresented clades. For the global model, the latter will be the prior. For an individual site η0 will be zero.

SmoothTree.CCD — Type

CCD(trees::Vector; [lmap=taxon_map])
CCD(trees::AbstractDict; [lmap=taxon_map])

A conditional clade distribution (CCD) object.

Input data can be either a vector of trees or a countmap of trees (see StatsBase.countmap).

Examples

julia> ccd = CCD([nw"((A,B),C);", nw"((B,C),A);", nw"((A,B),C);"])
CCD{UInt16}(n=3, Γ=7)

SmoothTree.EPABC — Type

EPABC(data, prior; kwargs...)

Expectation-propagation approximate Bayesian computation (EP-ABC, or likelihood free EP) algorithm struct. See ep! and pep!.

References

Barthelmé, Simon, and Nicolas Chopin. "Expectation propagation for likelihood-free inference." Journal of the American Statistical Association 109.505 (2014): 315-333.

Barthelmé, Simon, Nicolas Chopin, and Vincent Cottet. "Divide and conquer in ABC: Expectation-propagation algorithms for likelihood-free inference." Handbook of Approximate Bayesian Computation. Chapman and Hall/CRC, 2018. 415-434.

SmoothTree.MSC — Type

MSC(species_tree, initialization::Dict)

Multi-species coalescent distribution over gene trees for a given species tree with branch lengths in coalescent units.

The initialization is a dictionary recording for each leaf node of the species tree a vector of genes, which should be clade labels (i.e. powers of 2) (see taxonmap and the example below).

Example

julia> tree = nw"((F,(G,H):1):1,((B,((A,C):1,(D,E):1):1):1,O):1);";
       m = taxonmap(tree);
       init = Dict(id(n)=>[m[name(n)]] for n in getleaves(tree))
Dict{UInt16, Vector{UInt16}} with 9 entries:
  0x0005 => [0x0002]
  0x000d => [0x0020]
  0x0006 => [0x0004]
  0x000f => [0x0040]
  0x0010 => [0x0080]
  0x0009 => [0x0008]
  0x000c => [0x0010]
  0x0011 => [0x0100]
  0x0003 => [0x0001]

julia> M = SmoothTree.MSC(tree, init);
       randtree(M, m)
((F,(D,E)),((G,H),((B,(A,C)),O)));

SmoothTree.MSCModel — Type

MSCModel

An object for conducting variational species tree inference under the multispecies coalescent model. Note that:

We approximate the tree topology posterior by a MBM

We approximate the branch parameter posterior by independent Gaussians, one for each clade, i.e. representing the parameter for the branch leading to that clade as a crown group.

SmoothTree.MomBranchModel — Type

MomBranchModel

BranchModel in moment parameter space, not really used, more for convenience (interpretability).

SmoothTree.MomMBM — Type

MomMBM

A MBM model in moment parameter space.

SmoothTree.MomMBM — Method

MomMBM(x::CCD, β::BetaSplitTree, α::Real)
NatMBM(...)

Get the posterior mean Markov branching model assuming a Dirichlet prior distribution with parameter proportional to a Beta-splitting model and weight α (i.e. the Dirichlet parameter vector for each Categorical split distribution sums to α, so that α serves as a total pseudocount) and observed splits recorded in x.

SmoothTree.NatMBM — Type

NatMBM

A MBM model in natural parameter space.

SmoothTree.SparseSplits — Type

SparseSplits

SmoothTree.SparseSplits — Method

Get a natural parameter SparseSplits object from split counts d, assuming the Beta-splitting Dirichlet prior with pseudo-count α and shape parameter β.